service enhancing validations

+++ src/main/java/com/mensal/sliceCtrl/DTO/IngredientesDTO.java
@@ -4,6 +4,7 @@ import com.mensal.sliceCtrl.entity.AbstractEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Table;
+import jakarta.validation.constraints.DecimalMin;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
@@ -20,6 +21,6 @@ public class IngredientesDTO extends AbstractEntityDTO {
     private String nomeIngrediente;
 
     @NotNull(message = "É obrigatorio informar a quantidade de ingredientes")
-    private double qtdeIngrediente;
+    private int qtdeIngrediente;
 
 }
